backend/requirements.txt
fastapi==0.109.0
pydantic==2.6.0
firebase-admin==6.4.0
google-cloud-secret-manager==2.17.0